requirements.txt
# 汇通财经黄金资讯爬虫依赖包
requests>=2.25.1
beautifulsoup4>=4.9.3
lxml>=4.6.3

# 大模型API调用
openai>=1.0.0

# 数据处理和分析
pandas>=1.3.0
numpy>=1.21.0

# Web应用框架
streamlit>=1.28.0

# 数据可视化
plotly>=5.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
wordcloud>=1.9.0

# 配置管理
python-dotenv>=0.19.0

# 工具库
tqdm>=4.62.0
click>=8.0.0